Moved various traits to relationships directory
