+++ CMakeLists.txt
@@ -70,9 +70,6 @@ set(headers
   include/bit/stl/traits/composition/size_constant.hpp
   include/bit/stl/traits/composition/type_list.hpp
   include/bit/stl/traits/composition/void_t.hpp
-  include/bit/stl/traits/properties/arity.hpp
-  include/bit/stl/traits/properties/function_argument.hpp
-  include/bit/stl/traits/properties/index_of_type.hpp
   include/bit/stl/traits/properties/is_assignable.hpp
   include/bit/stl/traits/properties/is_const_member_function_pointer.hpp
   include/bit/stl/traits/properties/is_copy_assignable.hpp
@@ -100,8 +97,9 @@ set(headers
   include/bit/stl/traits/properties/is_trivially_move_constructible.hpp
   include/bit/stl/traits/properties/is_trivially_destructible.hpp
   include/bit/stl/traits/properties/is_volatile_member_function_pointer.hpp
-  include/bit/stl/traits/properties/nth_type.hpp
-  include/bit/stl/traits/properties/pointer_rank.hpp
+  include/bit/stl/traits/relationships/arity.hpp
+  include/bit/stl/traits/relationships/function_argument.hpp
+  include/bit/stl/traits/relationships/index_of_type.hpp
   include/bit/stl/traits/relationships/is_base_of_all.hpp
   include/bit/stl/traits/relationships/is_base_of_any.hpp
   include/bit/stl/traits/relationships/is_derived_from_all.hpp
@@ -114,6 +112,9 @@ set(headers
   include/bit/stl/traits/relationships/is_same_decay.hpp
   include/bit/stl/traits/relationships/is_same_sign.hpp
   include/bit/stl/traits/relationships/is_same_volatility.hpp
+  include/bit/stl/traits/relationships/nth_type.hpp
+  include/bit/stl/traits/relationships/resolves_overload.hpp
+  include/bit/stl/traits/relationships/pointer_rank.hpp
   include/bit/stl/traits/transformations/base_type.hpp
   include/bit/stl/traits/transformations/function_signature.hpp
   include/bit/stl/traits/transformations/match_cv_qualifiers.hpp
@@ -121,7 +122,6 @@ set(headers
   include/bit/stl/traits/transformations/match_sign_qualifiers.hpp
   include/bit/stl/traits/transformations/normalize_char.hpp
   include/bit/stl/traits/transformations/remove_pointers.hpp
-  include/bit/stl/traits/transformations/resolves_overload.hpp
 
   # Concepts
   include/bit/stl/concepts/AssociativeContainer.hpp
