+++ include/bit/stl/traits/relationships/function_argument.hpp
@@ -6,8 +6,8 @@
  *
  * \author Matthew Rodusek (matthew.rodusek@gmail.com)
  */
-#ifndef BIT_STL_TRAITS_PROPERTIES_FUNCTION_ARGUMENT_TYPE_HPP
-#define BIT_STL_TRAITS_PROPERTIES_FUNCTION_ARGUMENT_TYPE_HPP
+#ifndef BIT_STL_TRAITS_RELATIONSHIPS_FUNCTION_ARGUMENT_TYPE_HPP
+#define BIT_STL_TRAITS_RELATIONSHIPS_FUNCTION_ARGUMENT_TYPE_HPP
 
 #include "nth_type.hpp"
 
@@ -71,4 +71,4 @@ namespace bit {
   } // namespace stl
 } // namespace bit
 
-#endif /* BIT_STL_TRAITS_PROPERTIES_FUNCTION_ARGUMENT_TYPE_HPP */
+#endif /* BIT_STL_TRAITS_RELATIONSHIPS_FUNCTION_ARGUMENT_TYPE_HPP */
