+++ include/bit/stl/traits/relationships/index_of_type.hpp
@@ -6,8 +6,8 @@
  *
  * \author Matthew Rodusek (matthew.rodusek@gmail.com)
  */
-#ifndef BIT_STL_TRAITS_PROPERTIES_INDEX_OF_TYPE_HPP
-#define BIT_STL_TRAITS_PROPERTIES_INDEX_OF_TYPE_HPP
+#ifndef BIT_STL_TRAITS_RELATIONSHIPS_INDEX_OF_TYPE_HPP
+#define BIT_STL_TRAITS_RELATIONSHIPS_INDEX_OF_TYPE_HPP
 
 #include "../composition/size_constant.hpp"
 
@@ -42,4 +42,4 @@ namespace bit {
   } // namespace stl
 } // namespace bit
 
-#endif /* BIT_STL_TRAITS_PROPERTIES_INDEX_OF_TYPE_HPP */
+#endif /* BIT_STL_TRAITS_RELATIONSHIPS_INDEX_OF_TYPE_HPP */
