+++ include/bit/stl/traits/relationships/nth_type.hpp
@@ -6,8 +6,8 @@
  *
  * \author Matthew Rodusek (matthew.rodusek@gmail.com)
  */
-#ifndef BIT_STL_TRAITS_PROPERTIES_NTH_TYPE_HPP
-#define BIT_STL_TRAITS_PROPERTIES_NTH_TYPE_HPP
+#ifndef BIT_STL_TRAITS_RELATIONSHIPS_NTH_TYPE_HPP
+#define BIT_STL_TRAITS_RELATIONSHIPS_NTH_TYPE_HPP
 
 #include "../composition/identity.hpp"
 
@@ -35,4 +35,4 @@ namespace bit {
   } // namespace stl
 } // namespace bit
 
-#endif /* BIT_STL_TRAITS_PROPERTIES_NTH_TYPE_HPP */
+#endif /* BIT_STL_TRAITS_RELATIONSHIPS_NTH_TYPE_HPP */
