+++ include/bit/stl/traits/relationships/pointer_rank.hpp
@@ -6,8 +6,8 @@
  *
  * \author Matthew Rodusek (matthew.rodusek@gmail.com)
  */
-#ifndef BIT_STL_TRAITS_PROPERTIES_POINTER_RANK_HPP
-#define BIT_STL_TRAITS_PROPERTIES_POINTER_RANK_HPP
+#ifndef BIT_STL_TRAITS_RELATIONSHIPS_POINTER_RANK_HPP
+#define BIT_STL_TRAITS_RELATIONSHIPS_POINTER_RANK_HPP
 
 #include "../composition/size_constant.hpp"
 
@@ -42,4 +42,4 @@ namespace bit {
 
 
 
-#endif /* BIT_STL_TRAITS_PROPERTIES_POINTER_RANK_HPP */
+#endif /* BIT_STL_TRAITS_RELATIONSHIPS_POINTER_RANK_HPP */
